separate build and source sets
