+++ lib/wrapture.rb
@@ -24,10 +24,11 @@ require 'wrapture/c_source'
 require 'wrapture/c_to_cpp'
 require 'wrapture/c_to_cpp_wrapper'
 require 'wrapture/c_to_python'
+require 'wrapture/class_spec'
 require 'wrapture/comment'
 require 'wrapture/constant_spec'
 require 'wrapture/constants'
-require 'wrapture/class_spec'
+require 'wrapture/cpp_source'
 require 'wrapture/enum_spec'
 require 'wrapture/errors'
 require 'wrapture/function_spec'
@@ -37,6 +38,7 @@ require 'wrapture/rule_spec'
 require 'wrapture/param_spec'
 require 'wrapture/scope'
 require 'wrapture/source_file'
+require 'wrapture/source_set'
 require 'wrapture/struct_spec'
 require 'wrapture/template_spec'
 require 'wrapture/type_spec'
