+++ lib/wrapture/c_source.rb
@@ -23,6 +23,7 @@ require 'wrapture/c_source/c_if'
 require 'wrapture/c_source/c_include'
 require 'wrapture/c_source/c_pointer'
 require 'wrapture/c_source/c_source_file'
+require 'wrapture/c_source/c_source_set'
 require 'wrapture/c_source/c_struct'
 require 'wrapture/c_source/c_type'
 require 'wrapture/c_source/plain_c_block'
