+++ sig/build.rbs
@@ -17,9 +17,6 @@
 module Wrapture
   module Build
     def self.from_hash: (spec_hash) -> (Wrapture::Build::CmakeBuild | Wrapture::Build::PyprojectBuild)
-    def []: (String | Pathname) -> Wrapture::SourceFile
-    def include?: (String | Pathname) -> bool
-    def save: (?(String | Pathname)) -> Enumerable[Pathname]
-    def save_build_system: (?(String | Pathname)) -> Enumerable[Pathname]
+    def save_build_sources: (?(String | Pathname)) -> Enumerable[Pathname]
   end
 end
