+++ test/fixtures/cmake_c_library/build.yml
@@ -15,7 +15,7 @@
 # limitations under the License.
 
 build_system: "cmake"
-c_build:
+c_sources:
   name: "cmakeclib"
   lib_headers:
     - "cmakeclib.h"
